Add support for custom languages
