+++ language.js
@@ -1,72 +1,11 @@
 // Specify languages to translate to here
 // For now we can categories as replace, list and control
 // But we should look at another structure in the future.
+const php = require('./languages/php');
+const handlebars = require('./languages/handlebars');
 
-const php = {
-	replace: {
-		format: `<?php echo htmlspecialchars( $data[ "||%1||" ], ENT_QUOTES ); ?>`,
-	},
-	list: {
-		open: `<?php foreach ( $data[ "||%1||" ] as $item ) { ?>`,
-		close: `<?php } ?>`,
-		formatObject: `<?php echo htmlspecialchars( $item[ "||%1||" ], ENT_QUOTES ); ?>`,
-		formatPrimitive: `<?php echo htmlspecialchars( $item, ENT_QUOTES ); ?>`,
-	},
-	control: {
-		ifTruthy: {
-			open: `<?php if ( $data[ "||%1||" ] ) { ?>`,
-			close: '<?php } ?>',
-		},
-		ifFalsy: {
-			open: `<?php if ( ! $data[ "||%1||" ] ) { ?>`,
-			close: '<?php } ?>',
-		},
-		ifEqual: {
-			open: `<?php if ( $data[ "||%1||" ] === ||%2|| ) { ?>`,
-			close: '<?php } ?>',
-		},
-		ifNotEqual: {
-			open: `<?php if ( $data[ "||%1||" ] !== ||%2|| ) { ?>`,
-			close: '<?php } ?>',
-		},
-	}
-};
-
-const handlebars = {
-	replace: {
-		format: `{{||%1||}}`,
-	},
-	list: {
-		open: '{{#||%1||}}',
-		close: '{{/||%1||}}',
-		formatObject: `{{||%1||}}`,
-		formatPrimitive: `{{.}}`,
-	},
-	control: {
-		ifTruthy: {
-			open: '{{#if_truthy ||%1||}}',
-			close: '{{/if_truthy}}',
-		},
-		ifFalsy: {
-			open: '{{#if_falsy ||%1||}}',
-			close: '{{/if_falsy}}',
-		},
-		ifEqual: {
-			open: '{{#if_equal ||%1|| ||%2||}}',
-			close: '{{/if_equal}}',
-		},
-		ifNotEqual: {
-			open: '{{#if_not_equal ||%1|| ||%2||}}',
-			close: '{{/if_not_equal}}',
-		},
-	}
-};
-
-const languages = {
-	php,
-	handlebars,
-};
-
+// The main language object containing all registered languages.
+const languages = {};
 
 /**
  * Replaces tokens such as ||%1|| and ||%2|| with the arguments passed in.
@@ -103,7 +42,6 @@ function getLanguageString( language, type, targetString = [], argsArray = [] ) 
 	return createLanguageString( languageWithPath, argsArray );
 }
 
-
 function getLanguageReplace( language, target, subject ) {
 	return getLanguageString( language, 'replace', [ target ], [ subject ] );
 }
@@ -116,10 +54,18 @@ function getLanguageControl( language, targets, subjects ) {
 	return getLanguageString( language, 'control', targets, subjects );
 }
 
+function registerLanguage( language ) {
+	languages[ language.name ] = language;
+}
+
+// Now register the built-in languages.
+registerLanguage( php );
+registerLanguage( handlebars );
 
 module.exports = {
 	getLanguageReplace,
 	getLanguageList,
 	getLanguageControl,
+	registerLanguage,
 	languages,
 };
